+++ model/scratch/dffml_model_scratch/model/slr.py
